pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "psycopg>=3.3.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "validators>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
